requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
datefinder==0.7.0
dateparser==0.7.1
feedparser==5.2.1
goose3==3.1.6
idna==2.8
jieba==0.39
lxml==4.3.3
nltk>=3.4.5
numpy==1.16.3
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
regex==2019.8.19
requests==2.21.0
schedule==0.6.0
six==1.12.0
soupsieve==1.9.1
tzlocal==1.5.1
urllib3==1.24.3
requests[security]
